Show correct emulator URL (http://localhost:8090) in the console output
It used to be a specific IP address of a local machine. That was fine for older emulator versions, but not for the latest (11+).
The reason is that the IP address on HTTP-only origin does not allow to use some Web Browser features. Like Cache API.
So the emulator page was throwing error and sos.onReady() is never resolved.

related signageos/cli#28 signageos/cli#29